Move IntelligentScissors class and update worker
